requirements.txt
pandas
requests
python-dotenv
google-cloud-bigquery
pyspark==3.5.1
